Remove session state
With the main menu now in place, we no longer need to preserve
the initial grid state in user session. Instead, we force the
user to go through the menu, ensuring that when the live router
navigates to GameLive, its socket is already connected and so
mount() will only be called once.